backend/app/requirements.txt
sqlalchemy==1.4.27
psycopg2-binary==2.9.2
fastapi==0.70.0
pydantic==1.8.2
pydantic[email]==1.8.2
uvicorn==0.15.0
alembic==1.7.5
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0